[RENAME] 폴더 구조 정리 및 경로 수정
