+++ src/components/shared/Dropdown.jsx
@@ -1,7 +1,7 @@
 import { Link } from 'react-router-dom';
 import styled, { css } from 'styled-components';
-import productsList from 'assets/productsList.svg';
-import bookmark from 'assets/bookmark.svg';
+import productsList from 'assets/icons/productsList.svg';
+import bookmark from 'assets/icons/bookmark.svg';
 
 function Dropdown() {
   return (
