+++ src/components/shared/index.js
@@ -1,0 +1,27 @@
+import BookmarkButton from './BookmarkButton';
+import Dropdown from './Dropdown';
+import DropdownButton from './DropdownButton';
+import Filters from './Filters';
+import Footer from './Footer';
+import Header from './Header';
+import Logo from './Logo';
+import ModalPortal from './ModalPortal';
+import Product from './Product';
+import ProductModal from './ProductModal';
+import Products from './Products';
+import Toasts from './Toasts';
+
+export {
+  BookmarkButton,
+  Dropdown,
+  DropdownButton,
+  Filters,
+  Footer,
+  Header,
+  Logo,
+  ModalPortal,
+  Product,
+  ProductModal,
+  Products,
+  Toasts,
+};
