+++ src/pages/Main.jsx
@@ -1,10 +1,7 @@
 import { useMemo } from 'react';
 import { useSelector } from 'react-redux';
 import styled from 'styled-components';
-import Products from 'components/Products';
-import Toasts from 'components/Toasts';
-import ModalPortal from 'components/ModalPortal';
-import ProductModal from 'components/ProductModal';
+import { Products, Toasts, ModalPortal, ProductModal } from 'components';
 import sliceProducts from 'utils/functions/sliceProducts';
 
 function Main() {
