+++ src/pages/index.js
@@ -1,0 +1,5 @@
+import Bookmark from './Bookmark';
+import Main from './Main';
+import ProductsList from './ProductsList';
+
+export { Bookmark, Main, ProductsList };
