+++ src/utils/functions/index.js
@@ -1,0 +1,5 @@
+import filterProducts from './filterProducts';
+import setInitialFilter from './setInitialFilter';
+import sliceProducts from './sliceProducts';
+
+export { filterProducts, setInitialFilter, sliceProducts };
